Google login provider: Revoke app authorization
signOut() disconnect the user but keeps the app authorization binded to the user google account.
"revokeAuth" method calls "disconnect" and gives the user the ability to revoke the authorization from it's google account. In this way he won't need to log into his google account manager and manually revoking it.