Fix build to enable Angular v9 support
The current library on npm contains both `@angular/common`
and `rxjs`  in the distributable files. Apart from meaning that
these files are vastly larger than they need to be it also causes
the Angular `ngcc` compilation to fail, which prevents this
library from being used in an app compiled with Angular ivy.

This commit adds `rxjs` and `@angular/common` as externals
to the rollup configuration which solves this problem.