feat(config): Add "lang" field with language code value to "useValue" property
Description:
This commit introduces a new feature to enhance language selection for the Google login button in "@abacritt/angularx-social-login". The "lang" field, with a language code value, has been added to the "useValue" property in the configuration file. This addition allows developers to manually specify the language for the Google button interface, improving usability for multilingual applications.

Changes Made:

Added "lang" field to the "useValue" property in the configuration file
Enabled manual language selection for the Google login button
Related Issue: #762 - "Add manual support for language selection in the Google button"